api/requirements.txt
SQLAlchemy==2.0.29
Flask==3.0.3
Flask-Cors==4.0.0
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.3.0
segment-anything==1.0
torch==2.3.0
torchvision==0.18.0
matplotlib==3.8.4
tdqm==0.0.1
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1